+++ src/main/java/com/handelsbanken/ecommerce/checkout_service/domain/model/WatchCatalogue.java
@@ -4,10 +4,12 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.handelsbanken.ecommerce.checkout_service.domain.discount.DiscountStrategy;
 
+import java.math.BigDecimal;
+
 public record WatchCatalogue(
         String id,
         String name,
-        int unitPrice,
+        BigDecimal unitPrice,
         @JsonIgnore
         DiscountStrategy discountStrategy
 ) {
